authentication complete. routes protected

+++ client/src/components/WorkoutDetails.jsx
@@ -1,12 +1,20 @@
 import React from 'react'
 import { useWorkoutContext } from '../hooks/useWorkoutContext'
+import { useAuthContext } from '../hooks/useAuthContext';
 
 const WorkoutDetails = ({ workout}) => {
 
     const { dispatch } = useWorkoutContext();
+    const { user } = useAuthContext()
     const handleClick = async () => {
+
+        if (!user) { return }
+
         const response = await fetch('api/workouts/' + workout._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
